+++ CHANGELOG.md
@@ -1,13 +1,15 @@
 ## Release 2.0.0
 
 - Add support for tasks.
+- Rename the run paramater to autorun.
 - Change structured fact format from array to hash.
 - Fully qualify the path to the test command.
 - Normalize the waitforcert and user parameters of the puppet device command.
 
 ### Summary
 
-This release adds support for tasks.
+This release adds support for tasks, and uses Puppet Development Kit (PDK).
+.
 
 ## Release 1.1.1